Track downloads as pageviews (#1209)
* Track bitstream downloads with a different action type

using setActionUrl instead of setDownloadUrl means these appear as
TYPE=1 and not TYPE=3 in the matomo database.
This is not just for backwards compatiblity, there's https://github.com/matomo-org/developer-documentation/issues/735
which documents that the action_name is used only with page view
actions.

Additionally this commit cleans up some redundant parameters - they'd be
set the same way for all tracking requests.

* Assert that custom dimension contains the item handle